+++ app/src/main/kotlin/com/pitchedapps/frost/web/FrostWebViewClients.kt
@@ -71,6 +71,7 @@ open class FrostWebViewClient(val webCore: FrostWebViewCore) : BaseWebViewClient
             view.jsInject(
                     CssAssets.ROUND_ICONS.maybe(Prefs.showRoundedIcons),
                     CssHider.HEADER,
+                    CssHider.CORE,
                     CssHider.PEOPLE_YOU_MAY_KNOW.maybe(!Prefs.showSuggestedFriends && IS_FROST_PRO),
                     Prefs.themeInjector,
                     CssHider.NON_RECENT.maybe(webCore.url?.contains("?sk=h_chr") ?: false))