+++ docs/Changelog.md
@@ -6,6 +6,7 @@
 * Allow custom ringtones for notifications and messages (notification setting)
 * Improve logout logic when account is invalid
 * Fix material light background for certain items
+* Fix initial login not sticking
 
 ## v1.4.7
 * Update secondary background for transparent themes to be more visible.